+++ Assets/Editor/PxEnumDrawer.cs
@@ -14,6 +14,11 @@ namespace PseudoExtensibleEnum
     public class PxEnumDrawer : PropertyDrawer
     {
 
+        public override float GetPropertyHeight(SerializedProperty property, GUIContent label)
+        {
+            return EditorGUI.GetPropertyHeight(property, label);
+        }
+
         public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
         {
             PxEnumPropertyAttribute pxEnumPropertyAttribute = (PxEnumPropertyAttribute)attribute;